refactor: Movel all business logic from controller to service refactor: Adjust rabbitmq settings
