+++ packages/auth-ldap/package.json
@@ -27,7 +27,7 @@
     "@nestjs/config": "^1.0.1",
     "@nestjs/core": "^7.6.15",
     "@nestjs/jwt": "^7.2.0",
-    "@nestjs/microservices": "^7.6.15",
+    "@nestjs/microservices": "^7.6.17",
     "@nestjs/passport": "^7.1.5",
     "@nestjs/platform-express": "^7.6.15",
     "@prisma/client": "^2.23.0",
