+++ packages/auth-ldap/src/auth/auth.controller.ts
@@ -6,29 +6,16 @@ import {
   RmqContext,
   RpcException,
 } from '@nestjs/microservices';
-import { JwtService } from '@nestjs/jwt';
-import { UserService } from '../user/user.service';
-import { User } from '.prisma/client';
 import { AuthService } from './auth.service';
 import { Credentials } from './interfaces/credentials.interface';
-import { UserLdap } from './interfaces/user-ldap.interface';
 import { UserToken } from './interfaces/user-token.interface';
-import {
-  MSG_AUTHENTICATE,
-  MSG_VALIDATE,
-  ERR_AUTHENTICATE,
-  ERR_VALIDATE,
-} from './constants';
+import { MSG_AUTHENTICATE, MSG_VALIDATE } from './constants';
 
 @Controller('auth')
 export class AuthController {
   private readonly logger = new Logger(AuthController.name);
 
-  constructor(
-    private readonly userService: UserService,
-    private readonly authService: AuthService,
-    private readonly jwtService: JwtService
-  ) {}
+  constructor(private readonly authService: AuthService) {}
 
   @MessagePattern(MSG_AUTHENTICATE)
   async authenticate(
@@ -39,42 +26,10 @@ export class AuthController {
     const originalMsg = context.getMessage();
 
     try {
-      const user: UserLdap = await this.authService.getLDAPUser(credentials);
-
-      const userBd = await this.userService.getUserByUsername(user.uid);
-
-      let updatedUser: User;
-
-      if (userBd) {
-        updatedUser = await this.userService.updateUser({
-          where: { username: user.uid },
-          data: {
-            username: user.uid,
-            email: user.mail,
-            name: user.givenName,
-            surname: user.sn,
-          },
-        });
-      } else {
-        updatedUser = await this.userService.createUser({
-          username: user.uid,
-          email: user.mail,
-          name: user.givenName,
-          surname: user.sn,
-        });
-      }
-
-      const payload = {
-        username: updatedUser.username,
-        id: updatedUser.id,
-      };
-
-      return {
-        user: updatedUser,
-        access_token: this.jwtService.sign(payload),
-      };
+      const user = await this.authService.authenticate(credentials);
+      return user;
     } catch (err) {
-      throw new RpcException(ERR_AUTHENTICATE);
+      throw new RpcException(err);
     } finally {
       await channel.ack(originalMsg);
     }
@@ -90,10 +45,9 @@ export class AuthController {
 
     try {
       const tokenValid = this.authService.validateToken(token);
-
       return tokenValid;
     } catch (err) {
-      throw new RpcException(ERR_VALIDATE);
+      throw new RpcException(err);
     } finally {
       await channel.ack(originalMsg);
     }
